Create separate helpers file
