+++ music_assistant/providers/ai_radio/provider.py
@@ -18,7 +18,8 @@ from .constants import (
     DEFAULT_MAX_CONCURRENT_RUNS,
     SUPPORTED_FEATURES,
 )
-from .models import SessionState, coerce_int, utc_now_iso
+from .helpers import coerce_int, utc_now_iso
+from .models import SessionState
 from .runtime import AIRadioRuntimeMixin
 from .storage import AIRadioStorageMixin
